Home screen Refactor done Pending is Swipping tiles error
