--- lib/feature/create_business_card.dart/controller/LinearProgressIndicator_controller.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-
-class LinearProgressController extends GetxController {
-  final RxInt currentPage = 0.obs;
-  final PageController pageController = PageController();
-
-  void onPageChanged(int index) {
-    currentPage.value = index;
-  }
-
-  void nextPage() {
-    if (currentPage.value < 7) {
-      pageController.nextPage(
-        duration: const Duration(milliseconds: 500),
-        curve: Curves.ease,
-      );
-    }
-  }
-}
