+++ pubspec.yaml
@@ -29,20 +29,13 @@ environment:
 # versions available, run `flutter pub outdated`.
 dependencies:
   animate_do: any
-  barcode_scanner: ^3.6.2
   bottom_bar_matu: ^1.3.1
   cupertino_icons: ^1.0.2
   dotted_border: ^2.1.0
   flutter:
     sdk: flutter
   flutter_screenutil: ^5.9.0
-  get: any
-  iconly: ^1.0.1
   image_picker: ^1.0.5
-  provider: ^6.1.1
-  sizer: ^2.0.15
-  qrscan: ^0.3.3
-  flutter_calendar_carousel: ^2.4.2
 
 dev_dependencies:
   flutter_lints: ^2.0.0
